fix: guard the NTLM challenge payload fields against wrapping offsets
Client.Authenticate validated TargetName and TargetInfo as
int(BufferOffset + uint32(Len)). Both parts come from the server, and the sum
is taken in uint32: an offset of 0xFFFFFFFF with a length of 1 wraps to 0, so
the guard reads len(cmsg) < 0, clears, and the slice below it runs from
0xFFFFFFFF and panics.

Same defect class as the response decoders in this branch, on a path that is
just as exposed. The CHALLENGE_MESSAGE is parsed during session setup, before
any key is established, so it is reachable by an on-path attacker rather than
only by a malicious server.

Take both sums in uint64. The int() conversion would additionally go negative
on a 32-bit build.

internal/ntlm/server.go has the same pattern at four sites. It is dead code --
nothing in the repo constructs an ntlm.Server -- so it is left alone with the
request.go decoders, on the same reasoning.

+++ internal/ntlm/client.go
@@ -89,11 +89,14 @@ func (c *Client) Authenticate(cmsg []byte) (amsg []byte, err error) {
 	if targetNameMaxLen < targetNameLen {
 		return nil, errors.New("invalid target name format")
 	}
-	targetNameBufferOffset := le.Uint32(cmsg[16:20]) // cmsg.TargetNameBufferOffset
-	if len(cmsg) < int(targetNameBufferOffset+uint32(targetNameLen)) {
+	// The sum is taken in uint64: both fields are server-controlled, and in
+	// uint32 an offset near the type maximum wraps it below len(cmsg), which
+	// clears the guard for a slice that is out of range.
+	targetNameBufferOffset := uint64(le.Uint32(cmsg[16:20])) // cmsg.TargetNameBufferOffset
+	if uint64(len(cmsg)) < targetNameBufferOffset+uint64(targetNameLen) {
 		return nil, errors.New("invalid target name format")
 	}
-	targetName := cmsg[targetNameBufferOffset : targetNameBufferOffset+uint32(targetNameLen)] // cmsg.TargetName
+	targetName := cmsg[targetNameBufferOffset : targetNameBufferOffset+uint64(targetNameLen)] // cmsg.TargetName
 
 	if flags&NTLMSSP_NEGOTIATE_TARGET_INFO == 0 {
 		return nil, errors.New("invalid negotiate flags")
@@ -104,11 +107,11 @@ func (c *Client) Authenticate(cmsg []byte) (amsg []byte, err error) {
 	if targetInfoMaxLen < targetInfoLen {
 		return nil, errors.New("invalid target info format")
 	}
-	targetInfoBufferOffset := le.Uint32(cmsg[44:48]) // cmsg.TargetInfoBufferOffset
-	if len(cmsg) < int(targetInfoBufferOffset+uint32(targetInfoLen)) {
+	targetInfoBufferOffset := uint64(le.Uint32(cmsg[44:48])) // cmsg.TargetInfoBufferOffset
+	if uint64(len(cmsg)) < targetInfoBufferOffset+uint64(targetInfoLen) {
 		return nil, errors.New("invalid target info format")
 	}
-	targetInfo := cmsg[targetInfoBufferOffset : targetInfoBufferOffset+uint32(targetInfoLen)] // cmsg.TargetInfo
+	targetInfo := cmsg[targetInfoBufferOffset : targetInfoBufferOffset+uint64(targetInfoLen)] // cmsg.TargetInfo
 	info := newTargetInfoEncoder(targetInfo, utf16le.EncodeStringToBytes(c.TargetSPN))
 	if info == nil {
 		return nil, errors.New("invalid target info format")
